+++ spec/controllers/admin/feedbacks_controller_spec.rb
@@ -1,0 +1,34 @@
+# frozen_string_literal: true
+
+describe Admin::FeedbacksController do
+  describe 'GET#index' do
+    it_behaves_like 'required auth'
+    it_behaves_like 'required admin auth'
+
+    context 'for logged admins' do
+      sign_in_admin
+
+      context 'without companies' do
+        it 'renders index page' do
+          do_request
+
+          expect(response).to render_template :index
+        end
+      end
+
+      context 'with companies' do
+        before { create :feedback }
+
+        it 'renders index page' do
+          do_request
+
+          expect(response).to render_template :index
+        end
+      end
+    end
+
+    def do_request
+      get :index
+    end
+  end
+end